+++ src/components/alert/interfaces/index.ts
@@ -1,4 +1,5 @@
 export type {
+  AlertProps,
   AlertRootProps,
   AlertIconProps,
   AlertTitleProps,