+++ src/components/avatar/interfaces/index.ts
@@ -1,4 +1,5 @@
 export type {
+  AvatarProps,
   AvatarRootProps,
   AvatarImageProps,
   AvatarFallbackProps,